chore: combine scene 5 and 6
